fix(auth): move useSearchParams to client and wrap with Suspense
